third_party/blink/web_tests/animations/svg-use-keyframes-resolution.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
  @keyframes fade {
    from { opacity: 1; }
    to { opacity: 0; }
  }
  rect {
    animation: fade 100s paused;
  }
</style>
<svg>
  <defs>
    <rect id="target" width="100" height="100" fill="green" />
  </defs>
  <use href="#target" />
</svg>
<script>
  test(() => {
    // Force style recalc.
    document.body.offsetTop;

    // The <use> element creates a clone of #target inside a UA shadow root.
    const use = document.querySelector('use');
    const shadowRoot = use.shadowRoot || internals.shadowRoot(use);

    const clone = shadowRoot.querySelector('rect');
    assert_not_equals(clone, null, 'Clone rect should exist in shadow root');

    // The clone should still resolve @keyframes defined in the document scope.
    const animations = clone.getAnimations();
    assert_equals(animations.length, 1, 'Clone should have one animation');
    assert_equals(animations[0].animationName, 'fade',
        'Animation should resolve the "fade" keyframes');
  }, 'SVG <use> clone resolves @keyframes from document scope');
</script>
